fix next/link element - minor code destructuring - move custom menu data list to constants folder
