+++ src/components/menu/index.tsx
@@ -1,4 +1,6 @@
 import React from 'react';
+import { useRouter } from 'next/router';
+import Link from 'next/link';
 
 type Item = {
     icon: string | React.ReactElement;
@@ -7,11 +9,14 @@ type Item = {
 };
 
 type MenuItems = {
-    currentPage: string;
     items: Item[];
 };
 
-function Menu({ currentPage, items }: MenuItems) {
+function Menu({ items }: MenuItems) {
+    // routing page for active state
+    const router = useRouter();
+    const currentPage = router.pathname;
+
     return (
         <aside
             id="default-sidebar"
@@ -20,22 +25,21 @@ function Menu({ currentPage, items }: MenuItems) {
         >
             <div className="h-full px-3 py-4 overflow-y-auto bg-gray-50 dark:bg-gray-800">
                 <ul className="space-y-2 font-medium">
-                    {items.map((item, id) => (
+                    {items.map(({ path, icon, label }, id) => (
                         <li key={id}>
-                            <a
-                                href={item.path}
+                            <Link
+                                href={path}
                                 className={` flex items-center p-2 rounded-lg ${
-                                    item.path == currentPage
+                                    path == currentPage
                                         ? 'text-menu-secondary-color bg-menu-primary-color'
                                         : 'text-menu-primary-color '
                                 }   hover:bg-menu-primary-color  hover:text-menu-secondary-color`}
                             >
                                 <div className="w-6 h-6  transition duration-75 dark:text-gray-400 group-hover:text-menu-secondary-color dark:group-hover:text-white">
-                                    {item.icon}
+                                    {icon}
                                 </div>
-
-                                <span className="ml-3">{item.label}</span>
-                            </a>
+                                <span className="ml-3">{label}</span>
+                            </Link>
                         </li>
                     ))}
                 </ul>
